+++ Source/Modules/Catalog/Gapotchenko.FX.Memory/MemoryEqualityComparer.Equatable.cs
@@ -5,6 +5,9 @@
 // File introduced by: Oleksiy Gapotchenko
 // Year of introduction: 2020
 
+using System.IO.Hashing;
+using System.Runtime.InteropServices;
+
 namespace Gapotchenko.FX.Memory;
 
 partial class MemoryEqualityComparer
@@ -15,11 +18,14 @@ partial class MemoryEqualityComparer
 
         public override int GetHashCode(ReadOnlyMemory<T> obj)
         {
-            // FNV-1a
-            uint hash = 2166136261;
+            var hash = new XxHash3();
+            Span<int> buffer = stackalloc int[1];
             foreach (var i in obj.Span)
-                hash = (hash ^ (uint)(i?.GetHashCode() ?? 0)) * 16777619;
-            return (int)hash;
+            {
+                buffer[0] = i?.GetHashCode() ?? 0;
+                hash.Append(MemoryMarshal.AsBytes(buffer));
+            }
+            return (int)hash.GetCurrentHashAsUInt64();
         }
     }
 }