+++ Documentation/Changelog.md
@@ -6,8 +6,9 @@
 
 Release date: not released yet
 
-- Improved compatibility with AOT compilation
 - Added polyfills for `LeadingZeroCount(UInt32)`, `LeadingZeroCount(UInt64)`, and `Log2(UInt64)` methods of `BitOperations` class
+- Improved compatibility with AOT compilation
+- Improved hashing performance of `ArrayEqualityComparer`, `SpanEqualityComparer` and `MemoryEqualityComparer`. This is achieved by utilizing hardware-accelerated xxHash3 hash function instead of FNV-1a 
 - Fixed issue with `BitOperations.Log2(UInt32)` polyfill method that returned a non-zero value for zero argument on x64 processor architecture
 
 ### Gapotchenko.FX 2026.8